+++ BetterBulldozer/Patches/BulldozeToolSystemInitializeRaycastPatch.cs
@@ -32,8 +32,10 @@ namespace Better_Bulldozer.Patches
                 toolRaycastSystem.typeMask = betterBulldozerUISystem.MarkersFilter;
                 if ((betterBulldozerUISystem.MarkersFilter & TypeMask.Net) == TypeMask.Net)
                 {
-                    toolRaycastSystem.netLayerMask = Layer.MarkerPathway | Layer.MarkerTaxiway;
+                    toolRaycastSystem.netLayerMask = Layer.MarkerPathway | Layer.MarkerTaxiway | Layer.PowerlineLow | Layer.PowerlineHigh | Layer.WaterPipe | Layer.SewagePipe;
                     toolRaycastSystem.raycastFlags = RaycastFlags.Markers;
+                    toolRaycastSystem.utilityTypeMask = UtilityTypes.LowVoltageLine | UtilityTypes.HighVoltageLine | UtilityTypes.SewagePipe | UtilityTypes.SewagePipe;
+                    toolRaycastSystem.collisionMask = CollisionMask.OnGround | CollisionMask.Underground | CollisionMask.Overground;
                 }
                 else
                 {
@@ -50,6 +52,7 @@ namespace Better_Bulldozer.Patches
             {
                 toolRaycastSystem.typeMask = TypeMask.Net;
                 toolRaycastSystem.netLayerMask = Layer.Fence | Layer.LaneEditor;
+                toolRaycastSystem.raycastFlags |= RaycastFlags.Markers | RaycastFlags.EditorContainers;
             }
         }
     }