+++ BetterBulldozer/UI/src/mods/betterBulldozerSections/betterBulldozerSections.tsx
@@ -4,21 +4,37 @@ import { bindValue, trigger, useValue } from "cs2/api";
 import { VanillaComponentResolver } from "../VanillaComponentResolver/VanillaComponentResolver";
 import mod from "../../../mod.json";
 import spacesSrc from "./AreaToolBetterBulldozer.svg";
-import bypassConfirmationSrc from "./BypassQuestionmarkBB.svg";
-import gameplayManipulationSrc from "./CubeSimulationBB.svg";
 import staticObjectMarkersSrc from "./MarkersIconBetterBulldozer.svg";
-import lanesSrc from "./NetworkBB.svg";
-import networkMarkersSrc from "./DottedLinesMarkersBB.svg";
 import surfacesSrc from "./SurfaceIconBetterBulldozer.svg";
 import { tool } from "cs2/bindings";
 
 // These establishes the binding with C# side. Without C# side game ui will crash.
-// export const bulldozeToolActive$ = bindValue<boolean>(mod.id, 'BulldozeToolActive');
 export const gameplayManipulation$ = bindValue<boolean>(mod.id, 'GameplayManipulation');
 export const bypassConfirmation$ = bindValue<boolean>(mod.id, 'BypassConfirmation');
 export const raycastTarget$ = bindValue<number>(mod.id, 'RaycastTarget');
 export const areasFilter$ = bindValue<number>(mod.id, 'AreasFilter');
 export const markersFilter$ = bindValue<number>(mod.id, 'MarkersFilter');
+
+// These contain the coui paths to Unified Icon Library svg assets
+export const couiStandard =                         "coui://uil/Standard/";
+export const gameplayManipulationSrc =         couiStandard +  "CubeSimulation.svg";
+export const bypassConfirmationSrc =           couiStandard +  "BypassQuestionmark.svg";
+export const lanesSrc =                         couiStandard + "Network.svg";
+export const networkMarkersSrc =                couiStandard + "DottedLinesMarkers.svg";
+
+
+// Saving strings for events and translations.
+export const surfacesID =              "SurfacesFilterButton";
+export const spacesID =                "SpacesFilterButton";
+export const staticObjectsID =         "StaticObjectsFilterButton";
+export const networksFilterID =        "NetworksFilterButton";
+export const gameplayManipulationID =  "GameplayManipulationButton";
+export const bypassConfirmationID =    "BypassConfirmationButton";
+export const raycastMarkersID =        "RaycastMarkersButton";
+export const raycastAreasID =          "RaycastAreasButton";
+export const lanesID =                 "RaycastLanesButton";
+export const tooltipDescriptionPrefix ="YY_BETTER_BULLDOZER_DESCRIPTION.";
+export const sectionTitlePrefix =      "YY_BETTER_BULLDOZER.";
 
 // This functions trigger an event on C# side and C# designates the method to implement.
 export function handleClick(eventName: string) {
@@ -39,19 +55,6 @@ export const BetterBulldozerComponent: ModuleRegistryExtend = (Component : any) 
         const areasFilter = useValue(areasFilter$);
         const markersFilter = useValue(markersFilter$);
         
-        // Saving strings for events and translations.
-        const surfacesID =              "SurfacesFilterButton";
-        const spacesID =                "SpacesFilterButton";
-        const staticObjectsID =         "StaticObjectsFilterButton";
-        const networksFilterID =        "NetworksFilterButton";
-        const gameplayManipulationID =  "GameplayManipulationButton";
-        const bypassConfirmationID =    "BypassConfirmationButton";
-        const raycastMarkersID =        "RaycastMarkersButton";
-        const raycastAreasID =          "RaycastAreasButton";
-        const lanesID =                 "RaycastLanesButton";
-        const tooltipDescriptionPrefix ="YY_BETTER_BULLDOZER_DESCRIPTION.";
-        const sectionTitlePrefix =      "YY_BETTER_BULLDOZER.";
-
         // translation handling. Translates using locale keys that are defined in C# or fallback string here.
         const { translate } = useLocalization();
         const filterSectionTitle =          translate(sectionTitlePrefix + "Filter",                        "Filter");
@@ -80,7 +83,7 @@ export const BetterBulldozerComponent: ModuleRegistryExtend = (Component : any) 
         // It is important that we coordinate how to handle the tool options panel because it is possibile to create a mod that works for your mod but prevents others from doing the same thing.
         // If bulldoze tool active add better bulldozer sections.
         if (bulldozeToolActive) {
-            result.props.children?.unshift(
+            result.props.children?.push(
                 /* 
                 Adds a section for filters if raycasting areas or markers. Each of those sections has two buttons.
                 Adds a section for tool mode with 4 buttons.
@@ -90,26 +93,26 @@ export const BetterBulldozerComponent: ModuleRegistryExtend = (Component : any) 
                     { raycastingAreas ?
                         // This section is only showing if Raycasting areas. It includes filters for surfaces and spaces.
                         <VanillaComponentResolver.instance.Section title={filterSectionTitle}>
-                            <VanillaComponentResolver.instance.ToolButton className={VanillaComponentResolver.instance.toolButtonTheme.button} selected={surfacesFilter}    tooltip={surfacesFilterTooltip} onSelect={() => handleClick(surfacesID)}    src={surfacesSrc}></VanillaComponentResolver.instance.ToolButton>
-                            <VanillaComponentResolver.instance.ToolButton className={VanillaComponentResolver.instance.toolButtonTheme.button} selected={spacesFilter}      tooltip={spacesFilterTooltip}   onSelect={() => handleClick(spacesID)}      src={spacesSrc}></VanillaComponentResolver.instance.ToolButton>
+                                <VanillaComponentResolver.instance.ToolButton className={VanillaComponentResolver.instance.toolButtonTheme.button} selected={surfacesFilter}    tooltip={surfacesFilterTooltip} onSelect={() => handleClick(surfacesID)}  focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}  src={surfacesSrc}></VanillaComponentResolver.instance.ToolButton>
+                                <VanillaComponentResolver.instance.ToolButton className={VanillaComponentResolver.instance.toolButtonTheme.button} selected={spacesFilter}      tooltip={spacesFilterTooltip}   onSelect={() => handleClick(spacesID)}    focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}  src={spacesSrc}></VanillaComponentResolver.instance.ToolButton>
                         </VanillaComponentResolver.instance.Section>
                         : <></>
                     }
                     { raycastingMarkers ? 
                         // This section is only showing if Raycasting markers. It includes filters for static objects and networks.
                         <VanillaComponentResolver.instance.Section title={filterSectionTitle}>
-                            <VanillaComponentResolver.instance.ToolButton className={VanillaComponentResolver.instance.toolButtonTheme.button} selected={staticObjectMarkersFilter} tooltip={staticObjectMarkersTooltip}    onSelect={() => handleClick(staticObjectsID)}   src={staticObjectMarkersSrc}></VanillaComponentResolver.instance.ToolButton>
-                            <VanillaComponentResolver.instance.ToolButton className={VanillaComponentResolver.instance.toolButtonTheme.button} selected={markerNetworkFilter}       tooltip={markerNetworkTooltip}          onSelect={() => handleClick(networksFilterID)}  src={networkMarkersSrc}></VanillaComponentResolver.instance.ToolButton>
+                                <VanillaComponentResolver.instance.ToolButton className={VanillaComponentResolver.instance.toolButtonTheme.button} selected={staticObjectMarkersFilter} tooltip={staticObjectMarkersTooltip}    onSelect={() => handleClick(staticObjectsID)}   focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     src={staticObjectMarkersSrc}></VanillaComponentResolver.instance.ToolButton>
+                                <VanillaComponentResolver.instance.ToolButton className={VanillaComponentResolver.instance.toolButtonTheme.button} selected={markerNetworkFilter}       tooltip={markerNetworkTooltip}          onSelect={() => handleClick(networksFilterID)}  focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     src={networkMarkersSrc}></VanillaComponentResolver.instance.ToolButton>
                         </VanillaComponentResolver.instance.Section>
                         : <></>
                     }
-                     <VanillaComponentResolver.instance.Section title={toolModeTitle}>
-                        <VanillaComponentResolver.instance.ToolButton  selected={raycastingLanes}       tooltip={lanesTooltip}                  onSelect={() => handleClick(lanesID)}                   src={lanesSrc}                  className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
-                        <VanillaComponentResolver.instance.ToolButton  selected={gameplayManipulation}  tooltip={gameplayManipulationTooltip}   onSelect={() => handleClick(gameplayManipulationID)}    src={gameplayManipulationSrc}   className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
-                        <VanillaComponentResolver.instance.ToolButton  selected={bypassConfirmation}    tooltip={bypassConfirmationTooltip}     onSelect={() => handleClick(bypassConfirmationID)}      src={bypassConfirmationSrc}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
-                        <VanillaComponentResolver.instance.ToolButton  selected={raycastingMarkers}     tooltip={raycastMarkersTooltip}         onSelect={() => handleClick(raycastMarkersID)}          src={networkMarkersSrc}         className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
-                        <VanillaComponentResolver.instance.ToolButton  selected={raycastingAreas}       tooltip={raycastAreasTooltip}           onSelect={() => handleClick(raycastAreasID)}            src={surfacesSrc}               className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
-                     </VanillaComponentResolver.instance.Section>
+                    <VanillaComponentResolver.instance.Section title={toolModeTitle}>
+                            <VanillaComponentResolver.instance.ToolButton  selected={gameplayManipulation}  tooltip={gameplayManipulationTooltip}   onSelect={() => handleClick(gameplayManipulationID)}    src={gameplayManipulationSrc}  focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
+                            <VanillaComponentResolver.instance.ToolButton  selected={bypassConfirmation}    tooltip={bypassConfirmationTooltip}     onSelect={() => handleClick(bypassConfirmationID)}      src={bypassConfirmationSrc}    focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
+                            <VanillaComponentResolver.instance.ToolButton  selected={raycastingLanes}       tooltip={lanesTooltip}                  onSelect={() => handleClick(lanesID)}                   src={lanesSrc}                 focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
+                            <VanillaComponentResolver.instance.ToolButton  selected={raycastingMarkers}     tooltip={raycastMarkersTooltip}         onSelect={() => handleClick(raycastMarkersID)}          src={networkMarkersSrc}        focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
+                            <VanillaComponentResolver.instance.ToolButton  selected={raycastingAreas}       tooltip={raycastAreasTooltip}           onSelect={() => handleClick(raycastAreasID)}            src={surfacesSrc}              focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED}     className={VanillaComponentResolver.instance.toolButtonTheme.button}></VanillaComponentResolver.instance.ToolButton>
+                    </VanillaComponentResolver.instance.Section>
                 </>
             );
         }