+++ BetterBulldozer/ChangeLog.md
@@ -1,1 +1,4 @@
-﻿
+﻿# Patch V1.1.2
+* Added Dependency for Unified Icon Library.
+* Fixed Raycasting Lanes.
+* Added utilities networks to raycast marker networks. Also removes their floating icons with them.